refactor(mcp): make the client/server split real folder structure
src/mcp was flat despite the public API already naming ./mcp/client and
./mcp/server. The module now mirrors that boundary: client/ (manager,
connection, storage, catalog, invoker, rpc restore, runtime, transports,
OAuth provider, errors, x402), server/ (stateless entry, handlers, legacy
McpAgent, transports, event store, auth context, utils), with shared
types/rpc/abort and the compatibility barrel at the root. Files moved with
git mv so history follows; every module's exports are unchanged — public
import paths are identical and only build entry points and dist layout
moved (package.json exports updated in lockstep).
