+++ packages/agents/scripts/build.ts
@@ -15,10 +15,10 @@ const entries = [
   "src/chat/react.tsx",
   "src/chat-sdk/index.ts",
   "src/mcp/index.ts",
-  "src/mcp/client.ts",
-  "src/mcp/server.ts",
-  "src/mcp/do-oauth-client-provider.ts",
-  "src/mcp/x402.ts",
+  "src/mcp/client/index.ts",
+  "src/mcp/server/index.ts",
+  "src/mcp/client/do-oauth-client-provider.ts",
+  "src/mcp/client/x402.ts",
   "src/observability/index.ts",
   "src/observability/ai/index.ts",
   "src/schedules/index.ts",
