+++ packages/agents/src/mcp/client/catalog.ts
