+++ packages/agents/src/mcp/client/errors.ts
