+++ packages/agents/src/mcp/server/auth-context.ts
