+++ packages/agents/src/mcp/server/event-store.ts
