+++ packages/agents/src/mcp/server/handler-compat.ts
