+++ packages/agents/src/mcp/server/handler-errors.ts
