+++ packages/agents/src/mcp/server/handler-legacy-compat.ts
