+++ packages/agents/src/mcp/server/handler-legacy.ts
