+++ packages/agents/src/mcp/server/index.ts
