+++ packages/agents/src/mcp/server/sse-keepalive.ts
