+++ packages/agents/src/node-tests/mcp-server-bundle.test.ts
@@ -5,7 +5,7 @@ import { describe, expect, it } from "vitest";
 describe("agents/mcp/server bundle", () => {
   it("does not retain legacy, client, or Agent runtime modules", async () => {
     const result = await build({
-      entryPoints: [resolve("src/mcp/server.ts")],
+      entryPoints: [resolve("src/mcp/server/index.ts")],
       bundle: true,
       conditions: ["workerd", "browser", "import", "module"],
       external: ["cloudflare:*"],
@@ -22,8 +22,8 @@ describe("agents/mcp/server bundle", () => {
     expect(inputs).not.toContain("@modelcontextprotocol+client@");
     expect(inputs).not.toContain("partyserver");
     expect(inputs).not.toContain("partysocket");
-    expect(inputs).not.toContain("src/mcp/worker-transport.ts");
-    expect(inputs).not.toContain("src/mcp/handler-legacy.ts");
+    expect(inputs).not.toContain("src/mcp/server/worker-transport.ts");
+    expect(inputs).not.toContain("src/mcp/server/handler-legacy.ts");
     expect(inputs).not.toContain("src/mcp/index.ts");
   });
 });
