+++ packages/agents/src/tests/mcp/session-lifecycle.integration.test.ts
@@ -3,7 +3,7 @@ import { createExecutionContext } from "cloudflare:test";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import worker from "../worker";
 import type { MCPClientManager } from "../../mcp/client";
-import type { MCPServerRow } from "../../mcp/client-storage";
+import type { MCPServerRow } from "../../mcp/client/storage";
 import { type McpHarness, withMcpHarness } from "../capabilities/mcp-client";
 
 type RecordedRequest = {
